Move various functions to tpm_common.go
